Refactor src/utils, add `isNumberInRange`
